% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_read.R
\name{align_read}
\alias{align_read}
\title{Align a read to a reference sequence based on the CIGAR string}
\usage{
align_read(reference_seq, read_seq, cigar)
}
\arguments{
\item{reference_seq}{A character string representing the reference sequence.}

\item{read_seq}{A character string representing the read sequence.}

\item{cigar}{A character string representing the CIGAR string of the read.}
}
\value{
A list with two elements, the reference sequence and the
read sequence.
}
\description{
This function takes as input a reference sequence, a read sequence, and a
CIGAR string, and returns the aligned reference and read sequences.
}
\examples{
reference_seq <- "AGCTTAGCTAGCTACCTATATCTTGGTCTTG"
read_seq <- "AGCTTAGCTAGCTAC-TATCTTGGCCTTG"
cigar <- "14M1D11M1X6M"
align_read(reference_seq, read_seq, cigar) 

}
\seealso{
\code{\link{show_alignment}}
}
\author{
Anna Grattagliano\cr Politecnico di Milano\cr Maintainer:
Anna Grattagliano\cr E-Mail: \href{mailto:anna.grattagliano@mail.polimi.it}{anna.grattagliano@mail.polimi.it}
}
