% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_read.R
\name{extract_read}
\alias{extract_read}
\title{extract a specific read ID’s Information from a BAM/SAM file #'
This function loads a BAM/SAM file and a specific read ID and
will returnthe necessary information}
\usage{
extract_read(file, read_id)
}
\arguments{
\item{file}{the path to the BAM/SAM input file}

\item{read_id}{is the identifier of the read}
}
\value{
the read information regarding the specifico read ID
}
\description{
extract a specific read ID’s Information from a BAM/SAM file #'
This function loads a BAM/SAM file and a specific read ID and
will returnthe necessary information
}
\examples{
library(Rsamtools)
library(data.table)
file <- system.file("extdata", "ERR188273_chrX.bam", 
package = "AlignmentView")
read_id <- "ERR188273.11944385"
extract_read(file,read_id)

}
\seealso{
\code{\link{show_alignment}}
}
\author{
Anna Grattagliano\cr Politecnico di Milano\cr Maintainer:
Anna Grattagliano\cr E-Mail: \href{mailto:anna.grattagliano@mail.polimi.it}{anna.grattagliano@mail.polimi.it}
}
