% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_cigar.R
\name{parse_cigar}
\alias{parse_cigar}
\title{Parse a CIGAR string
This function takes as input the CIGAR string and finds the
alphabetic and numeric parts and seperates them}
\usage{
parse_cigar(cigar)
}
\arguments{
\item{cigar}{is the CIGAR string of the read}
}
\value{
a list with operations and their corresponding
lengths
}
\description{
Parse a CIGAR string
This function takes as input the CIGAR string and finds the
alphabetic and numeric parts and seperates them
}
\examples{
cigar <- "51M1D24M"
parse_cigar(cigar)

}
\seealso{
\code{\link{show_alignment}}
}
\author{
Anna Grattagliano\cr Politecnico di Milano\cr Maintainer:
Anna Grattagliano\cr E-Mail: \href{mailto:anna.grattagliano@mail.polimi.it}{anna.grattagliano@mail.polimi.it}
}
