% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_alignment.R
\name{show_alignment}
\alias{show_alignment}
\title{Show alignment of a read to the reference genome}
\usage{
show_alignment(file, read_id, reference_genome)
}
\arguments{
\item{file}{A character string representing the path to the file
containing read information.}

\item{read_id}{A character string representing the ID of the read
to be aligned.}

\item{reference_genome}{A \code{BSgenome} object representing
the reference genome.}
}
\value{
Prints the alignment of the read to the reference genome.
}
\description{
This function takes a file containing read information, a read ID,
and a reference genome,and prints the alignment of the specified
read to the reference genome.
}
\examples{
library(BSgenome.Hsapiens.UCSC.hg38)
file <- system.file("extdata", "ERR188273_chrX.bam", 
package = "AlignmentView")
read_id <- "ERR188273.11944385"
genome <- BSgenome.Hsapiens.UCSC.hg38
show_alignment(file, read_id, genome)

}
\seealso{
\code{\link{parse_cigar}}, \code{\link{align_read}},
\code{\link{colorize_sequence}}
}
